% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{diff_i}
\alias{diff_i}
\title{Differential interaction weighting}
\usage{
diff_i(x)
}
\arguments{
\item{x}{}
}
\description{
Set of differential network methods taken from O. Basha et al., Bioinformatics (2020).
}
