% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{diff_n}
\alias{diff_n}
\title{Differential node weighting}
\usage{
diff_n(g1, g2, name1 = "weight", name2 = "weight",
  nameOut = "weight")
}
\arguments{
\item{g1}{a node weighted igraph network}

\item{g2}{a node weighted igraph network}

\item{name1}{deafult 'weight', the name of the node attribute to be used for g1}

\item{name2}{deafult 'weight', the name of the node attribute to be used for g2}

\item{nameOut}{deafult 'weight', the name of the node attribute to be used for the otuput graph}
}
\value{
a network where the node weights reprsent the difference in node weights from the two input networks.
}
\description{
Set of differential network methods taken from O. Basha et al., Bioinformatics (2020).
}
\details{
Takes two networks that differ only in  node weights (canonically log2 expression).
\code{diff_n} calculates the weight \code{n} of the output network as the difference between the weight in network one to that in network two.
 In the case of a multigroup comparison (classes > 2) the second network node weights should be the median of each class's weights.
}
