% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-helpers.R
\name{.v2e_directed}
\alias{.v2e_directed}
\title{Incoming directed weights}
\usage{
.v2e_directed(g, w)
}
\arguments{
\item{g}{graph}

\item{w}{vertex weights}
}
\value{

}
\description{
internal for v2e
}
