% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-helpers.R
\name{.v2e_vMax}
\alias{.v2e_vMax}
\title{max of vertex weights over edges}
\usage{
.v2e_vMax(g, w)
}
\arguments{
\item{g}{graph}

\item{w}{vertex weights}
}
\value{

}
\description{
internal for v2e. Takes a graph and vertex weights as input.
For each edge in the graph the weight is defined as the sum of the
weights of the two adjacent vertices
}
