% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_score}
\alias{drug_score}
\title{drug potential scores}
\format{
A dataframe with 15781 rows and two columns:
\describe{
  \item{name}{uniprotswissprot}
  \item{druggability}{Fpocket drug score (0,1)}
}
}
\usage{
drug_score
}
\description{
drug potential scores
}
\keyword{datasets}
