% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{filterScores}
\alias{filterScores}
\title{Filter scores}
\usage{
filterScores(s, g)
}
\arguments{
\item{s}{A 2 column data-frame with \code{col_names = c("protein", "score")}}

\item{g}{An igrpah graph object}
}
\value{
A 2 column data-frame with \code{col_names = c("protein", "score")}.
  Where all proteins are present as nodes of \code{g}
}
\description{
Filter a set of scores (i.e. a 2 col dataframe) to remove scored
proteins that are not in our graph of interest.
}
