% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{geometricMean}
\alias{geometricMean}
\title{geometricMean}
\usage{
geometricMean(x)
}
\arguments{
\item{x}{numeric vector where all `x > 0`}
}
\value{

}
\description{
Calculate geometric mean. This is an internal for `rankTable` as such
there is no testing. it will break if `any(x < 1 | is.na(x))`
}
