% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{getDrugScore}
\alias{getDrugScore}
\title{getDrugScore}
\usage{
getDrugScore(v)
}
\arguments{
\item{v}{character vector of uniprot IDs}
}
\value{
vector of druggability scores, NA if the score is missing
}
\description{
given a vector of protein names return
the druggability scores as a vector
}
