% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-helpers.R
\name{getLCS}
\alias{getLCS}
\title{Get largest connected subgraph}
\usage{
getLCS(g, simplify_g = F)
}
\arguments{
\item{g}{an igraph graph object}
}
\value{
an igraph object
}
\description{
Given an igraph, g, return the largest connected subgraph.
}
