% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-helpers.R
\name{getSubnet}
\alias{getSubnet}
\title{Get Subnetwork}
\usage{
getSubnet(g, v)
}
\arguments{
\item{g}{an igraph graph object, with \code{$name} node attribute}

\item{v}{A character vector of node names}
}
\value{
An igraph object
}
\description{
Given an igraph graph and a set of nodes, return the induced subgraph
  with self and multiple edges removed
}
