% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnet.R
\name{hnet_prepare}
\alias{hnet_prepare}
\title{Hnet Prepare}
\usage{
hnet_prepare(g, out_dir = getwd())
}
\arguments{
\item{g}{The PPI graph to be used, as an igraph object}

\item{out_dir}{where to save the graph files, default is `getwd()`}
}
\description{
Prepare a PPI graph for use with Hierarchical Hotnet
}
