% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnet.R
\name{hnet_read_results}
\alias{hnet_read_results}
\title{Hnet read results}
\usage{
hnet_read_results(file, first_clust = TRUE)
}
\arguments{
\item{file}{Path to the results file}

\item{first_clust}{Logical, whether to return the first (largest) cluster
or all of them.}
}
\value{
If `first_clust == TRUE` then a vector containing the uniprot IDs of the largest cluster from the Hierarchical Hotnet

  If `first_clust == FALSE` then a list containing a vector of uniprot IDs for each cluster returned in descending order of size
}
\description{
Read the results file from Hierarchical Hotnet
}
