% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{huri_graph}
\alias{huri_graph}
\title{Human reference interactome version 3.0}
\format{
An igraph object with 8010 vertices and 50568 edges
}
\source{
\url{http://interactome.baderlab.org/} for further details details
}
\usage{
huri_graph
}
\description{
A PPI developed from large scale chemical genetic yeast 2 hybrid screening.
All interactions were validated in at least two screens.
}
\keyword{datasets}
