% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-helpers.R
\name{makeContCol}
\alias{makeContCol}
\title{colourGraphStat}
\usage{
makeContCol(g, nStatFunc = igraph::degree,
  pal = grDevices::colorRampPalette(viridisLite::viridis(5)),
  scale = FALSE)
}
\arguments{
\item{g}{the igraph graph you want to analyse}

\item{nStatFunc}{a function that returns a node level statistic. Default is
\code{\link[igraph]{degree}}}

\item{pal}{a palette function, should take an integer arguement and return
a character vector of colours. See \code{\link[grDevices]{colorRamp}}}

\item{scale}{Whether to scale the output of the nStatFunc function}
}
\value{
a set of colour codes to be assigned to V(g)$colour
}
\description{
Given a graph
}
