% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-helpers.R
\name{netStats}
\alias{netStats}
\title{Network statistics}
\usage{
netStats(g, norm = FALSE, named = NULL, weights = NULL)
}
\arguments{
\item{g}{the igraph graph you want to analyse}

\item{norm}{Whether to normalise the scores to between 0 and 1}

\item{named}{are graph vertices named? If true F-pocket drug score will be included
in the table}

\item{weights}{passed through to igraph functions for betweeness, closeness, and eigen centrality.
See \code{\link[igraph]{betweenness}} for details}
}
\value{
A data.frame with four or five columns: \code{c(degree, betweeness, closeness, eigen_centrality, maybe(drug_score))}.
  For details on what each of these represent see each of the respective \code{igraph} functions.
}
\description{
Given an \code{igraph} graph object return a table with vertex statistics.
If the graph is names (or `named` is `TRUE`) F-pocket druggability statistics
are returned too
}
\examples{
# a graph
grph <- igraph::barabasi.game(30, directed = FALSE)
igraph::V(grph)$name <- c(letters, LETTERS[1:4])
netStats(grph)
}
\seealso{
\code{\link[igraph]{degree}}, \code{\link[igraph]{betweenness}},
  \code{\link[igraph]{closeness}}, \code{\link[igraph]{eigen_centrality}}
}
