% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{pref_i}
\alias{pref_i}
\title{Preferntial edge weighting}
\usage{
pref_i(g1, g2, name1 = "weight", name2 = "weight", nameOut = "weight")
}
\arguments{
\item{g1}{a node weighted igraph network}

\item{g2}{a node weighted igraph network}

\item{name1}{deafult 'weight', the name of the node attribute to be used for g1}

\item{name2}{deafult 'weight', the name of the node attribute to be used for g2}

\item{nameOut}{deafult 'weight', the name of the edge attribute to be used for the otuput graph}
}
\value{

}
\description{
Set of differential network methods taken from O. Basha et al., Bioinformatics (2020).
}
\details{
Computed as the sum of the preferential scores (pref_n's)of the interacting nodes.
}
