% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{pref_n}
\alias{pref_n}
\title{Preferential node weighting}
\usage{
pref_n(g1, g2, name1 = "weight", name2 = "weight", nameOut = "weight")
}
\arguments{
\item{g1}{a node weighted igraph network}

\item{g2}{a node weighted igraph network}

\item{name1}{deafult 'weight', the name of the node attribute to be used for g1}

\item{name2}{deafult 'weight', the name of the node attribute to be used for g2}

\item{nameOut}{deafult 'weight', the name of the node attribute to be used for the otuput graph}
}
\value{

}
\description{
Set of differential network methods taken from O. Basha et al., Bioinformatics (2020).
\code{pref_n} calculate the preferential expression of a node as described by  A. R. Sonawane et al., Cell Reports. 21, 1077–1088 (2017).
}
\details{
The preferential score of a node is effectively the differential node weighting divided by the IQR of the score in g2
}
