% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{rankTable}
\alias{rankTable}
\title{rankTable}
\usage{
rankTable(x, nameCol = NULL, aggFUN = geometricMean,
  ties.method = "min")
}
\arguments{
\item{x}{a data.frame with numerical scores to be ranked}

\item{nameCol}{the column name containing row labels, or NULL if no labels.
Default = NULL}

\item{aggFUN}{the function used to aggregate ranks, see details. Default =
geometric mean}

\item{ties.method}{how tied ranks should be dealt with. See
\code{\link[matrixStats]{colRanks}}}
}
\value{
a data.frame with the same shape as the input. Each column is
  trasformed to ranks (max value gets max rank) and then each row aggregated
  under \code{aggRank} using \code{aggFUN}
}
\description{
Given a data.frame with numerical scores return the rank table. Scores are
converted to ranks and then aggregated using \code{aggFUN}
}
\details{
If \code{nameCol} is used to specify a column this will be held
  aside when ranks are calculated then re-added before the rank table is
  returned
}
