% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential.R
\name{strip_attr}
\alias{strip_attr}
\title{Strip graph attributes}
\usage{
strip_attr(g)
}
\arguments{
\item{g}{}
}
\value{
the graph g with al vertex attributes removed
}
\description{
removes all vertex and edge attributes from a graph.
internal function for prepping graphs to test if structures are the same
}
