% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-helpers.R
\name{v2e}
\alias{v2e}
\title{v2e}
\usage{
v2e(g, method, w = NULL)
}
\arguments{
\item{g}{an igraph graph object}

\item{method}{See details ~~to write~~}

\item{w}{Numerical vector of vertex weights}
}
\value{
the graph \code{g} with weighted edges
}
\description{
Vertex to edge: reloaded
}
\details{
Transform a vertex-weighted undirected graph to an edge weighted graph
}
