% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores.R
\name{z2o}
\alias{z2o}
\title{Zero to One: range scaling}
\usage{
z2o(x, incZero = TRUE)
}
\arguments{
\item{x}{Numeric vector to be scaled}

\item{incZero}{Logical: whether the bottom of the interval should be open or closed}
}
\value{
the rescaled vector x, the range will be [0,1] or (0,1) depending on the value of \code{incZero}
}
\description{
Utility function: converts a numeric vector to the range [0,1] or (0,1)
}
