% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ascher_graph}
\alias{ascher_graph}
\title{PPI generated by the Ascher Lab at the Baker Institute}
\format{
An igraph object with 15756 vertices and 174414 edges:
}
\source{
\email{david.ascher@unimelb.edu.au} for details
}
\usage{
ascher_graph
}
\description{
A PPI made from combining multiple online interaction networks to make a meta-graph of public interaction data.
Edges are supported by a range of different evidence types
}
\keyword{datasets}
