% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ascher_graph}
\alias{ascher_graph}
\title{PPI generated by the Ascher Lab at the Baker Institute}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}
}
\source{
\email{david.ascher@unimelb.edu.au} for details
}
\usage{
ascher_graph
}
\description{
A PPI: give some details about what the interactions represent here.
}
\keyword{datasets}
