% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-helpers.R
\name{.v2e_directed}
\alias{.v2e_directed}
\title{Incoming directed weights}
\usage{
.v2e_directed(g, w)
}
\arguments{
\item{g}{graph}

\item{w}{vertex weights}
}
\value{

}
\description{
internal for v2e. Takes a graph and vertex weights as input.
Each edge in the graph is replaced by mutual directed edges, that is:
the edge \code{A:-:B} becomes the edges \code{A:->:B} and \code{A:<-:B}
The weight of the vertex at the head of each edge then
becomes the edge weight.
}
