% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug_score}
\alias{drug_score}
\title{drug potential scores}
\format{\describe{
  \item{name}{uniprotswissprot}
  \item{druggability}{Fpocket}
}}
\usage{
drug_score
}
\description{
drug potential scores
}
\keyword{datasets}
