% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converters.R
\name{g2p}
\alias{g2p}
\title{Gene to Protein}
\usage{
g2p(g)
}
\arguments{
\item{g}{a character vector of human gene symbols}
}
\value{
a character vector of human uniprot ids
}
\description{
Given a vector of gene names return matching proteins names, or NA if no match
}
