% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-helpers.R
\name{getLCS}
\alias{getLCS}
\title{Get largest connected subgraph}
\usage{
getLCS(g, simplify_g = F)
}
\arguments{
\item{g}{an igraph graph object}

\item{simplify_g}{whether to remove self / multiple edges before returning
the graph, usually not neccessary.}
}
\value{
an igraph object
}
\description{
Given an igraph, g, return the largest connected subgraph.
}
\examples{
grph1 <- igraph::barabasi.game(26, directed = FALSE)
grph2 <- igraph::barabasi.game(20, directed = FALSE)


igraph::V(grph1)$name <- letters
igraph::V(grph2)$name <- LETTERS[1:20]

grph <- igraph::graph.union(grph1, grph2)

 plot(grph)

 plot(getLCS(grph))
}
