% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{huri_graph}
\alias{huri_graph}
\title{Human reference interactome version 3.0}
\format{A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://interactome.baderlab.org/} for further details details
}
\usage{
huri_graph
}
\description{
A PPI developed from large scale chemical genetic yeast 2 hybrid screening.
All interactions were validated in at least two screens.
}
\keyword{datasets}
