% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network-helpers.R
\name{netStats}
\alias{netStats}
\title{Network statistics}
\usage{
netStats(g, norm = FALSE)
}
\arguments{
\item{g}{the igraph graph you want to analyse}

\item{norm}{Whether to normalise the scores to between 0 and 1}
}
\value{
A data.frame with four columns: \code{c(degree, betweeness, closeness, eigen_centrality)}.
  For details on what each of these represent see each of the respective \code{igraph} functions.
}
\description{
Given an \code{igraph} graph object return a table with vertex statistics
}
\examples{
# a graph
grph <- igraph::barabasi.game(30, directed = FALSE)
netStats(grph)
}
\seealso{
\code{\link[igraph]{degree}}, \code{\link[igraph]{betweeness}},
  \code{\link[igraph]{closeness}}, \code{\link[igraph]{eigen_centrality}}
}
