% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{STRING_graph}
\alias{STRING_graph}
\title{The STRING PPI}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}
}
\source{
\url{https://string-db.org/} for details
}
\usage{
STRING_graph
}
\description{
This PPI is based on experimentally validated interactions withiin the STRING database.
Only interactions with a score > 400 (high-confidence) are included
}
\keyword{datasets}
